+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `class-factory` will be documented in this file.
 
+## v1.0.1 - 2022-09-03
+
+### Fixed
+- Make sure closures in definition get filled attributes array.
+
+### Updated
+- Improve clarity in `collapseStates()` method.
+
 ## v1.0.0 - 2022-08-30
 
 🏭